requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
blis==1.0.1
blurhash==1.1.4
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
huggingface-hub==0.25.2
icecream==2.1.3
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
Mastodon.py==1.8.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
networkx==3.4.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pytextrank==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
tokenizers==0.20.1
torch==2.5.0
torch-geometric==2.6.1
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.15.4
